Moved helper functions for IdTable from JoinTest to a seperate helper file.
