+++ test/EngineTest.cpp
@@ -21,43 +21,7 @@
 #include "util/SourceLocation.h"
 #include "./util/GTestHelpers.h"
 #include "engine/idTable/IdTable.h"
-
-ad_utility::AllocatorWithLimit<Id>& allocator() {
-  static ad_utility::AllocatorWithLimit<Id> a{
-      ad_utility::makeAllocationMemoryLeftThreadsafeObject(
-          std::numeric_limits<size_t>::max())};
-  return a;
-}
-
-auto I = [](const auto& id) {
-  return Id::makeFromVocabIndex(VocabIndex::make(id));
-};
-
-// For easier reading. We repeat that type combination so often, that this
-// will make things a lot easier in terms of reading and writing.
-using vectorTable = std::vector<std::vector<size_t>>;
-
-/*
- * Return an 'IdTable' with the given 'tableContent'. all rows must have the
- * same length.
-*/
-IdTable makeIdTableFromVector(vectorTable tableContent) {
-  AD_CHECK(!tableContent.empty());
-  IdTable result{tableContent[0].size(), allocator()};
-
-  // Copying the content into the table.
-  for (const auto& row: tableContent) {
-    AD_CHECK(row.size() == result.numColumns()); // All rows of an IdTable must have the same length.
-    const size_t backIndex{result.size()};
-    result.emplace_back();
-
-    for (size_t c = 0; c < row.size(); c++) {
-      result(backIndex, c) = I(row[c]);
-    }
-  }
-
-  return result;
-}
+#include "./util/IdTableHelpers.h"
 
 TEST(EngineTest, distinctTest) {
   IdTable inp{makeIdTableFromVector({{1, 1, 3, 7}, {6, 1, 3, 6}, {2, 2, 3, 5}, {3, 6, 5, 4}, {1, 6, 5, 1}})};
